+++ NEWS.md
@@ -1,3 +1,5 @@
+# ympes (development version)
+
 # ympes 1.3.0
 
 - New assertion function `assert_scalar_whole()`.