+++ NEWS.md
@@ -1,4 +1,4 @@
-# ympes (development version)
+# ympes 1.4.0
 
 - New function `grepvrows()` which is identical to `greprows()` except with the
   default `value = TRUE`.